Changed the datatype for bikeId in gps_data from smallint to int - this is reflected in code by a change from int to long.

+++ CityBikes/Library/AllBikesLocation.cs
@@ -15,7 +15,7 @@ namespace Library
         /// <param name="context">A database context from which data should be retrieved.</param>
         /// <param name="Id">The id of the bike to find the location of.</param>
         /// <returns>The GPSLocation of the bike with <paramref name="Id"/> or null if var location is null</returns>
-        public static GPSLocation GetBikeLocation(this Database context, int Id)
+        public static GPSLocation GetBikeLocation(this Database context, long Id)
         {
             var location = (from bike in context.gps_data where bike.bikeId == Id orderby bike.queried descending select bike);
 
@@ -27,7 +27,7 @@ namespace Library
         /// </summary>
         /// <param name="context">A database context from which data should be retrieved.</param>
         /// <returns>An IEnumerable containing a Tuple for each bike with its bikeId and its location </returns>
-        public static IEnumerable<Tuple<int, GPSLocation>> GetBikeLocations(this Database context)
+        public static IEnumerable<Tuple<long, GPSLocation>> GetBikeLocations(this Database context)
         {
             var latest = from bike in context.gps_data
                          group bike by bike.bikeId into b
