+++ CityBikes/Library/AvailableBikes.cs
@@ -16,10 +16,10 @@ namespace Library
         /// </summary>
         /// <param name="context">A database context from which data should be retrieved.</param>
         /// <returns>A collection of bikes and their location</returns>
-        public static IEnumerable<Tuple<int, GPSLocation>> GetAvailableBikes(this Database context)
+        public static IEnumerable<Tuple<long, GPSLocation>> GetAvailableBikes(this Database context)
         {
-            Dictionary<int, GPSLocation> positions = AllBikesLocation.GetBikeLocations(context).ToDictionary(x => x.Item1, x => x.Item2);
-            Dictionary<int, DateTime> immobile = BikeStandstill.GetBikesImmobile(context).ToDictionary(x => x.Item1, x => x.Item2);
+            Dictionary<long, GPSLocation> positions = AllBikesLocation.GetBikeLocations(context).ToDictionary(x => x.Item1, x => x.Item2);
+            Dictionary<long, DateTime> immobile = BikeStandstill.GetBikesImmobile(context).ToDictionary(x => x.Item1, x => x.Item2);
 
             var immobileTimeSpan = new TimeSpan(0, IMMOBILE_MINUTES, 0);
             DateTime now = DateTime.Now;
