+++ CityBikes/Library/BikeStandstill.cs
@@ -14,11 +14,11 @@ namespace Library
         /// </summary>
         /// <param name="context">A database context from which data should be retrieved.</param>
         /// <returns>A collection of bikes and park-times.</returns>
-        public static IEnumerable<Tuple<int, DateTime>> GetBikesImmobile(this Database context)
+        public static IEnumerable<Tuple<long, DateTime>> GetBikesImmobile(this Database context)
         {
-            Dictionary<int, gps_data> firstData = new Dictionary<int, gps_data>();
-            Dictionary<int, gps_data> lastData = new Dictionary<int, gps_data>();
-            List<int> stopped = new List<int>();
+            Dictionary<long, gps_data> firstData = new Dictionary<long, gps_data>();
+            Dictionary<long, gps_data> lastData = new Dictionary<long, gps_data>();
+            List<long> stopped = new List<long>();
 
             foreach (var v in from gps in context.gps_data orderby gps.queried select gps)
             {
@@ -48,7 +48,7 @@ namespace Library
         /// <param name="context">A database context from which data should be retrieved.</param>
         /// <param name="immobileSince">Any bikes that were parked after <paramref name="immobileSince"/> will not be returned.</param>
         /// <returns>A collection of bikes and park-times.</returns>
-        public static IEnumerable<Tuple<int, DateTime>> GetBikesImmobile(this Database context, DateTime immobileSince)
+        public static IEnumerable<Tuple<long, DateTime>> GetBikesImmobile(this Database context, DateTime immobileSince)
         {
             return GetBikesImmobile(context).Where(b => b.Item2 < immobileSince);
         }
